spec: Factor MockStorageApi out of webstorage.spec.js
This is useful elsewhere.
